ts3 updates needed for ts3 setup1 for ITL. Will work with both setups. These are minor compatibility tweaks.

+++ BNL_T03/ppump_acq/v0/ccseoppump.py
@@ -96,7 +96,7 @@ try:
         if ((len(tokens) > 0) and (tokens[0] == 'ppump')):
     
             exptime = float(tokens[1])
-            imcount = float(tokens[2])
+            imcount = int(tokens[2])
             nshifts  = float(tokens[3])
 
             result = arcsub.synchCommand(10,"setHeader","SequenceNumber",seq)
